Remove unused variables to eliminate build warnings
Remove unused variable 'x' from pi-cfs-bug-repro.c:
- run_normal_1: variable completely unused
- run_normal_2: used in busy-wait, replaced with comment
- run_deadline: used in busy-wait, replaced with comment

Fix gettid() format warnings in dl_getparam.c:
- Cast gettid() return to long and use %ld format
- gettid() returns pid_t which can vary in size

Build is now completely warning-free.

+++ tools/dl_getparam.c
@@ -162,7 +162,7 @@ int main (int argc, char *argv[])
 		exit(EXIT_FAILURE);
 	}
 
-	printf("main thread at %llu [%d]\n", timespec_to_nsec(&now), gettid());
+	printf("main thread at %llu [%ld]\n", timespec_to_nsec(&now), (long)gettid());
 
 	pthread_create(&thread, NULL, run_deadline, NULL);
 
@@ -179,7 +179,7 @@ int main (int argc, char *argv[])
 		exit(EXIT_FAILURE);
 	}
 
-	printf("main dies at %llu [%d]\n", timespec_to_nsec(&now), gettid());
+	printf("main dies at %llu [%ld]\n", timespec_to_nsec(&now), (long)gettid());
 
 	return 0;
 }
